timeIsInterval {spacetime}	R Documentation

retrieve, or set, information whether time reflects instance (FALSE) or intervals (TRUE)

Description

retrieve, or set, information whether time reflects instance (FALSE) or intervals (TRUE)

Usage

 
timeIsInterval(x, ...)
timeIsInterval(x) <- value

Arguments

x
object, of any class

...
ignored

value
logical; sets the timeIsInterval value

Value

logical; this function sets or retrieves the attribute timeIsInterval of x, UNLESS x is of class ST, in which case it sets or retrieves this attribute for the time slot of the object, i.e. timeIsInterval(x@time) <- value

Note

From spacetime 0.8-0 on, timeIsInterval is dropped in favour of a more generic time intervals by specifying endTime of each observation

See Also

over, timeIsInterval

[Package spacetime version 1.3-3 Index]