nbMult {spacetime}	R Documentation

convert a spatial nb object to a matching STF object

Description

convert a spatial nb object to a matching STF object

Usage

 
nbMult(nb, st, addT = TRUE, addST = FALSE)

Arguments

nb
object of class nb (see package spdep), which is valid for the spatial slot of object st: length(nb) should equal length(st@sp)

st
object of class STF

addT
logical; should temporal neighbours be added?

addST
logical; should spatio-temporal neighbours be added?

Details

if both addT and addST are false, only spatial neighbours are added for each time replicate.

details are found in

Giovana M. de Espindola, Edzer Pebesma, Gilberto Câmara, 2011. Spatio-temporal regression models for deforestation in the Brazilian Amazon. STDM 2011, The International Symposium on Spatial-Temporal Analysis and Data Mining, University College London - 18th-20th July 2011.

Value

object of class nb

Author(s)

Edzer Pebesma

[Package spacetime version 1.3-3 Index]